install generate yaml
