+++ scripts/install.py
@@ -6,7 +6,7 @@ import shutil
 import getpass
 
 import datetime
-#from shared_workflow.load_config import load
+# from shared_workflow.load_config import load
 from shared_workflow import load_config as ldcfg
 import ConfigParser
 import argparse
@@ -14,9 +14,12 @@ from management import create_mgmt_db
 
 # TODO: namespacing
 from shared_workflow.shared import *
-print 
+from collections import OrderedDict
+from qcore import utils
+
+print
 workflow_config = ldcfg.load(os.path.dirname(os.path.realpath(__file__)), "workflow_config.json")
-workflow_root=workflow_config['gm_sim_workflow_root']
+workflow_root = workflow_config['gm_sim_workflow_root']
 global_root = workflow_config["global_root"]
 tools_dir = os.path.join(global_root, 'EMOD3D/tools')
 bin_process_dir = os.path.join(global_root, 'workflow/scripts')
@@ -36,10 +39,11 @@ stat_dir = os.path.join(global_root, 'StationInfo')
 latest_ll_dir = os.path.join(global_root, 'StationInfo/grid')
 latest_ll = 'non_uniform_with_real_stations_latest'
 
-#default values
-#TODO: after enabling different dt for LF and HF, this might need to change
+# default values
+# TODO: after enabling different dt for LF and HF, this might need to change
 default_dt = 0.005
 default_hf_dt = 0.005
+
 
 
 def q_accept_custom_rupmodel():
@@ -223,10 +227,9 @@ def q_final_confirm(run_name, yes_statcords, yes_model_params):
 
 def action(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_stoch_pairs, params_vel_path,
            stat_file_path, vs30_file_path, vs30ref_file_path, MODEL_LAT, MODEL_LON, MODEL_ROT, hh, nx, ny, nz, sufx,
-           sim_duration, flo, vel_mod_params_dir, yes_statcords, yes_model_params,dt = default_dt, hf_dt = default_hf_dt):
-    lf_sim_root_dir, hf_dir, bb_dir, figures_dir = os.path.join(sim_dir, "LF"), os.path.join(sim_dir,
-                                                                                             "HF"), os.path.join(
-        sim_dir, "BB"), os.path.join(sim_dir, "Figures")
+           sim_duration, flo, vel_mod_params_dir, yes_statcords, yes_model_params, dt=default_dt, hf_dt=default_hf_dt):
+
+    lf_sim_root_dir, hf_dir, bb_dir, figures_dir = os.path.join(sim_dir, "LF"), os.path.join(sim_dir, "HF"), os.path.join(sim_dir, "BB"), os.path.join(sim_dir, "Figures")
 
     dir_list = [sim_dir, lf_sim_root_dir, hf_dir, bb_dir, figures_dir]
     if not os.path.isdir(user_root):
@@ -245,7 +248,7 @@ def action(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_sto
     #    exe('ln -s %s/submit_emod3d.py %s'%(bin_process_dir,sim_dir))
     shutil.copy(os.path.join(workflow_root, "version"), sim_dir)
     shutil.copy(os.path.join(bin_process_dir, "submit.sh"), sim_dir)
-    
+
     # shutil.copy(os.path.join(bin_process_dir, "submit_emod3d.sh"), sim_dir)
     # shutil.copy(os.path.join(bin_process_dir, "submit_post_emod3d.sh"), sim_dir)
     # shutil.copy(os.path.join(bin_process_dir, "submit_hf.sh"), sim_dir)
@@ -254,7 +257,7 @@ def action(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_sto
     # shutil.copy(os.path.join(bin_process_dir, "params.py.template"), os.path.join(sim_dir, "params.py"))
 
     # Rename params.py.template to params.py
-    shutil.move(os.path.join(sim_dir, "params.py.template"), os.path.join(sim_dir, "params.py"))
+    #shutil.move(os.path.join(sim_dir, "params.py.template"), os.path.join(sim_dir, "params.py"))
 
     if not yes_model_params:
         print "Generation of model params has been skipped."
@@ -265,6 +268,7 @@ def action(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_sto
     bin_process_ver = "slurm"
 
     srf_files, stoch_files = zip(*srf_stoch_pairs)
+
     with open(os.path.join(sim_dir, "params_base.py"), "w") as f:
         f.write("run_name='%s'\n" % run_name)
         f.write("version='%s'\n" % emod3d_version)
@@ -312,7 +316,7 @@ def action(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_sto
         f.write("\n#dir for vel_mod \n")
         f.write("vel_mod_params_dir = '%s'\n" % vel_mod_params_dir)
         f.write("GRIDFILE = '%s' #gridout-x used to be referred to as GRIDFILE by gen_ts \n" % (
-        os.path.join(vel_mod_params_dir, 'gridfile%s' % sufx)))
+            os.path.join(vel_mod_params_dir, 'gridfile%s' % sufx)))
         f.write("GRIDOUT = '%s'\n" % (os.path.join(vel_mod_params_dir, 'gridout%s' % sufx)))
         f.write("#input for statgrid gen\n")
         f.write("MODEL_COORDS = '%s'\n" % os.path.join(vel_mod_params_dir, 'model_coords%s' % sufx))
@@ -351,7 +355,7 @@ def action(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_sto
             f.write("STAT_FILES=[stat_file]\n")
 
     print dir_list[0]
-    #set_permission(dir_list[
+    # set_permission(dir_list[
     #                   0])  # if user_root is first time created, recursively set permission from there. otherwise, set permission from sim_dir
 
     sys.path.append(sim_dir)
@@ -399,9 +403,136 @@ def action(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_sto
         print "Generation of statcords is skipped. You need to fix params_base.py manually"
 
 
+def create_fault_params_dict(sim_dir, event_name, run_name, run_dir, vel_mod_dir, srf_dir, srf_stoch_pairs, params_vel_path,
+           stat_file_path, vs30_file_path, vs30ref_file_path, MODEL_LAT, MODEL_LON, MODEL_ROT, hh, nx, ny, nz, sufx,
+           sim_duration, flo, vel_mod_params_dir, yes_statcords, yes_model_params, dt=default_dt, hf_dt=default_hf_dt):
+
+    lf_sim_root_dir = os.path.join(sim_dir, "LF")
+    hf_dir = os.path.join(sim_dir, "HF")
+    bb_dir = os.path.join(sim_dir, "BB")
+
+    dir_list = [sim_dir, lf_sim_root_dir, hf_dir, bb_dir]
+    if not os.path.isdir(user_root):
+        dir_list.insert(0, user_root)
+
+    verify_user_dirs(dir_list)
+
+    for filename in glob.glob(os.path.join(recipe_dir, '*.*')):
+        if filename == "README.md":
+            continue
+        shutil.copy(filename, sim_dir)
+
+        # TODO: the next two lines are two files for old post-processing
+        # shutil.copy(os.path.join(gmsa_dir,"parametersStation.py"),sim_dir)
+        # shutil.copy(os.path.join(gmsa_dir,"runPostProcessStation.ll"),sim_dir)
+    #    exe('ln -s %s/submit_emod3d.py %s'%(bin_process_dir,sim_dir))
+    shutil.copy(os.path.join(workflow_root, "version"), sim_dir)
+    shutil.copy(os.path.join(bin_process_dir, "submit.sh"), sim_dir)
+
+
+    if not yes_model_params:
+        print "Generation of model params has been skipped."
+        print "Re-directing related params to files under %s" % vel_mod_dir
+        vel_mod_params_dir = vel_mod_dir
+
+    # TODO: get rid of this
+    bin_process_ver = "slurm"
+
+    srf_files, stoch_files = zip(*srf_stoch_pairs)
+
+    fault_params_dict = OrderedDict()
+    fault_params_dict['run_name'] = run_name
+    fault_params_dict['version'] = emod3d_version
+    fault_params_dict['bin_process_ver'] = bin_process_ver
+    fault_params_dict['global_root'] = global_root
+    fault_params_dict['tools_dir'] = tools_dir
+    fault_params_dict['user_root'] = user_root
+    fault_params_dict['run_dir'] = run_dir
+    fault_params_dict['sim_dir'] = sim_dir
+    fault_params_dict['lf_sim_root_dir'] = lf_sim_root_dir
+    fault_params_dict['hf_dir'] = hf_dir
+    fault_params_dict['bb_dir'] = bb_dir
+    fault_params_dict['srf_dir'] = srf_dir
+    fault_params_dict['srf_file'] = ''
+    fault_params_dict['vel_mod_dir'] = vel_mod_dir
+    fault_params_dict['v_mod_1d_dir'] = v_mod_1d_dir
+    fault_params_dict['params_vel'] = params_vel_path
+    fault_params_dict['sim_duration'] = sim_duration
+    fault_params_dict['flo'] = flo
+
+    fault_params_dict['vm'] = OrderedDict()
+    fault_params_dict['vm']['MODEL_LAT'] = MODEL_LAT
+    fault_params_dict['vm']['MODEL_LON'] = MODEL_LON
+    fault_params_dict['vm']['MODEL_LON'] = MODEL_ROT
+    fault_params_dict['vm']['hh'] = hh
+    fault_params_dict['vm']['nx'] = nx
+    fault_params_dict['vm']['nz'] = nz
+    fault_params_dict['vm']['ny'] = ny
+    fault_params_dict['vm']['vel_mod_params_dir'] = vel_mod_params_dir
+    fault_params_dict['vm']['sufx'] = sufx
+    fault_params_dict['vm']['GRIDFILE'] = os.path.join(vel_mod_params_dir, 'gridfile%s' % sufx)
+    fault_params_dict['vm']['GRIDOUT'] = os.path.join(vel_mod_params_dir, 'gridout%s' % sufx)
+    fault_params_dict['vm']['MODEL_COORDS'] = os.path.join(vel_mod_params_dir, 'model_coords%s' % sufx)
+    fault_params_dict['vm']['MODEL_PARAMS'] = os.path.join(vel_mod_params_dir, 'model_params%s' % sufx)
+    fault_params_dict['vm']['MODEL_BOUNDS'] = os.path.join(vel_mod_params_dir, 'model_bounds%s' % sufx)
+
+    fault_params_dict['hf'] = OrderedDict()
+    fault_params_dict['hf']['hf_dt'] = hf_dt
+
+    fault_params_dict['emod3d'] = OrderedDict()
+    fault_params_dict['emod3d']['dt'] = dt
+
+    if stat_file_path == "":
+        # stat_path seems to empty, assigning all related value to latest_ll
+        print "stat_file_path is not specified."
+        print "Using %s" % latest_ll
+        run_stat_dir = os.path.join(stat_dir, event_name)
+        stat_file_path = os.path.join(run_stat_dir, event_name + '.ll')
+        vs30_file_path = os.path.join(run_stat_dir, event_name + '.vs30')
+        vs30ref_file_path = os.path.join(run_stat_dir, event_name + '.vs30ref')
+        # creating sub-folder for run_name
+        # check if folder already exist
+        if not os.path.isdir(run_stat_dir):
+            # folder not exist, creating
+            os.mkdir(run_stat_dir)
+            # making symbolic link to latest_ll
+            cmd = "ln -s %s %s" % (os.path.join(latest_ll_dir, latest_ll + '.ll'), stat_file_path)
+            exe(cmd)
+            # making symbolic link to lastest_ll.vs30 and .vs30ref
+            cmd = "ln -s %s %s" % (os.path.join(latest_ll_dir, latest_ll + '.vs30'), vs30_file_path)
+            exe(cmd)
+            cmd = "ln -s %s %s" % (os.path.join(latest_ll_dir, latest_ll + '.vs30ref'), vs30ref_file_path)
+            exe(cmd)
+    fault_params_dict['stat_vs_est'] = vs30_file_path
+    fault_params_dict['stat_vs_ref'] = vs30ref_file_path
+
+    if stat_file_path is not None:
+        fault_params_dict['stat_file'] = stat_file_path
+
+    # if yes_model_params:
+    #     print "Producing model params. It may take a minute or two"
+    #     from gen_coords import gen_coords
+    #     gen_coords()
+    #     print "Done"
+
+    if yes_statcords:
+        print "Producing statcords and FD_STATLIST. It may take a minute or two"
+
+        # Create Stat_cord & statList
+        import statlist2gp
+        fd_statcords, fd_statlist = statlist2gp.main(stat_file=stat_file_path)
+        print "Done"
+        fault_params_dict['stat_coords'] = fd_statcords
+        fault_params_dict['FD_STATLIST'] = fd_statlist
+    else:
+        print "Generation of statcords is skipped. You need to fix params_base.py manually"
+
+    return fault_params_dict
+
+
 def show_instruction(sim_dir):
     try:
-        print "Removing probably incomplete "+os.path.join(sim_dir, "params_base.pyc")
+        print "Removing probably incomplete " + os.path.join(sim_dir, "params_base.pyc")
         os.remove(os.path.join(sim_dir, "params_base.pyc"))
     except Exception, e:
         print e.args
@@ -567,13 +698,18 @@ def main_local():
 
     srf_files, ___ = zip(*srf_stoch_pairs)
     create_mgmt_db.create_mgmt_db([], sim_dir, srf_files=srf_files)
-    #saves the location of mgmt_db to params_base.py
-    with open(os.path.join(sim_dir,"params_base.py"),"a") as f:
+    # saves the location of mgmt_db to params_base.py
+    with open(os.path.join(sim_dir, "params_base.py"), "a") as f:
         f.write("mgmt_db_location='%s'\n" % sim_dir)
-
 
     print "Installation completed"
     show_instruction(sim_dir)
+
+    fault_params_dict = create_fault_params_dict(sim_dir, event_name, run_name, run_dir, vel_mod_dir_full, srf_dir, srf_stoch_pairs, params_vel_path,
+           stat_file_path, vs30_file_path, vs30ref_file_path, MODEL_LAT, MODEL_LON, MODEL_ROT, hh, nx, ny, nz, sufx,
+           sim_duration, flo, vel_mod_params_dir, yes_statcords, yes_model_params)
+    fault_params_dict.update['mgmt_db_location'] = sim_dir
+    utils.dump_yaml(fault_params_dict)
 
 
 def main_remote(cfg):
@@ -628,23 +764,23 @@ def main_remote(cfg):
 
 
 if __name__ == '__main__':
-    
-    parser = argparse.ArgumentParser()
-    
-    parser.add_argument('--user_root',type=str,default=None)
-    parser.add_argument('--sim_cfg',type=str,default=None)
-    
-    parser.add_argument('--srf_dir',type=str,default=None)
-    parser.add_argument('--vm_dir',type=str,default=None)
-    parser.add_argument('--v1d_dir',type=str,default=None)
-    parser.add_argument('--station_dir',type=str,default=None)
 
-#if user desire using specific files, they should call the action function directly instead
-#    parser.add_argument('--srf',type=str,default=None)
-#    parser.add_argument('--stoch',type=str,default=None)
-#    parser.add_argument('--vm',type=str,default=None)
-#    parser.add_argument('--station',type=str,default=None)
-    
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument('--user_root', type=str, default=None)
+    parser.add_argument('--sim_cfg', type=str, default=None)
+
+    parser.add_argument('--srf_dir', type=str, default=None)
+    parser.add_argument('--vm_dir', type=str, default=None)
+    parser.add_argument('--v1d_dir', type=str, default=None)
+    parser.add_argument('--station_dir', type=str, default=None)
+
+    # if user desire using specific files, they should call the action function directly instead
+    #    parser.add_argument('--srf',type=str,default=None)
+    #    parser.add_argument('--stoch',type=str,default=None)
+    #    parser.add_argument('--vm',type=str,default=None)
+    #    parser.add_argument('--station',type=str,default=None)
+
     args = parser.parse_args()
 
     # If the additional options provided, check if the folder exist
@@ -652,39 +788,39 @@ if __name__ == '__main__':
         path_to_check = getattr(args, arg)
         if path_to_check is not None:
             if not os.path.exists(path_to_check):
-                print "Error: path not exsist: %s"%path_to_check
+                print "Error: path not exsist: %s" % path_to_check
                 sys.exit()
             else:
-                print "%s is set to %s"%(arg,path_to_check)
+                print "%s is set to %s" % (arg, path_to_check)
         else:
             continue
 
-    #change corresponding variables to the args provided
-    
+    # change corresponding variables to the args provided
+
     if args.user_root != None:
-        #TODO:bad hack, fix this with parsing
+        # TODO:bad hack, fix this with parsing
         user_root = args.user_root
 
     if args.srf_dir != None:
-        #TODO:bad hack, fix this with parsing
+        # TODO:bad hack, fix this with parsing
         srf_default_dir = args.srf_dir
 
     if args.vm_dir != None:
-        #TODO:bad hack, fix this with parsing
+        # TODO:bad hack, fix this with parsing
         vel_mod_dir = args.vm_dir
-    
+
     if args.v1d_dir != None:
-        #TODO:bad hack, fix this with parsing
-        v_mod_1d_dir = args.v1d_dir 
+        # TODO:bad hack, fix this with parsing
+        v_mod_1d_dir = args.v1d_dir
 
     if args.station_dir != None:
-        #TODO:bad hack, fix this with parsing
+        # TODO:bad hack, fix this with parsing
         stat_dir = args.station_dir
- 
-    #if sim_cfg parsed, run main_remote(which has no selection)
+
+    # if sim_cfg parsed, run main_remote(which has no selection)
     if args.sim_cfg != None:
         cfg = args.sim_cfg
-        #check if the cfg exist, to prevent break
+        # check if the cfg exist, to prevent break
         if not os.path.exists(cfg):
             print "Error: No such file exists: %s" % cfg
             sys.exit()
@@ -692,13 +828,13 @@ if __name__ == '__main__':
             main_remote(cfg)
     else:
         main_local()
-    
-    #if len(sys.argv) == 1:
-    #    main_local()
-    #else:
-    #    cfg = sys.argv[1]
-    #    if not os.path.exists(cfg):
-    #        print "Error: No such file exists: %s" % cfg
-    #        sys.exit()
-    #    else:
-    #        main_remote(cfg)
+
+        # if len(sys.argv) == 1:
+        #    main_local()
+        # else:
+        #    cfg = sys.argv[1]
+        #    if not os.path.exists(cfg):
+        #        print "Error: No such file exists: %s" % cfg
+        #        sys.exit()
+        #    else:
+        #        main_remote(cfg)
